Move errors.py back to types package
