+++ src/cryptoserve/exercises/__init__.py
@@ -1,2 +1,1 @@
-from cryptoserve.exercises.errors import *
 from cryptoserve.exercises.simple_hash import simple_hash
