+++ src/cryptoserve/types/errors.py
